reafactor

+++ Server/Game/Interface/IEntity.cs
@@ -2,8 +2,8 @@ public interface IEntity
 {
     public string Id { get; init; }
     public bool Collision { get; set; }
-    public int PosX { get; set; }
-    public int PosY { get; set; }
+    public double PosX { get; set; }
+    public double PosY { get; set; }
     public int Width { get; set; }
     public int Height { get; set; }
     public bool CheckCollistion(IEntity entity);
